Move some new functionality from qa_utils to misc_utils. Update the changelog, changing it to be a .rst file, not a .md file.
