+++ dcicutils/misc_utils.py
@@ -815,6 +815,9 @@ class CachedField:
 
 
 def make_counter(start=0, step=1):
+    """
+    Creates a counter that generates values counting from a given start (default 0) by a given step (default 1).
+    """
     storage = [start]
 
     def counter():
@@ -823,3 +826,89 @@ def make_counter(start=0, step=1):
         return value
 
     return counter
+
+
+class UncustomizedInstance(Exception):
+    """
+    Reports a helpful error for access to a CustomizableProperty that has not been properly set.
+    """
+
+    def __init__(self, instance, *, field):
+        self.instance = instance
+        self.field = field
+        declaration_class, declaration = self._find_field_declared_class(instance, field)
+        self.declaration_class = declaration_class
+        context = ""
+        if declaration_class == 'instance':
+            context = " from instance"
+        elif declaration_class:
+            context = " from class %s" % full_object_name(declaration_class)
+        message = ("Attempt to access field %s%s."
+                   " It was expected to be given a custom value in a subclass: %s."
+                   % (field, context, declaration.description))
+        super().__init__(message)
+
+    @staticmethod
+    def _find_field_declared_class(instance, field):
+        instance_value = instance.__dict__.get(field)
+        is_class = isinstance(instance, type)
+        if instance_value:
+            return instance if is_class else 'instance', instance_value
+        else:
+            for cls in instance.__mro__ if is_class else instance.__class__.__mro__:
+                cls_value = cls.__dict__.get(field)
+                if cls_value:
+                    return cls, cls_value
+            raise RuntimeError("%s does not have a field %s." % (instance, field))
+
+
+class CustomizableProperty(property):
+    """
+    Declares a class variable to require customization. See help on getattr_customized for details.
+    """
+
+    def __init__(self, field, *, description):
+
+        self.field = field
+        self.description = description
+
+        def uncustomized(instance):
+            raise UncustomizedInstance(instance=instance, field=field)
+
+        super().__init__(uncustomized)
+
+    def __str__(self):
+        return "<%s %s>" % (self.__class__.__name__, self.field)
+
+
+def getattr_customized(thing, key):
+    """
+    Like getattr, but if the value is a CustomizableProperty, gives a helpful error explaining need to customize.
+
+    This avoids inscrutible errors or even dangerous confusions that happen when an abstract class requires setting
+    of variables that the user of the class might forget to set or not realize they're supposed to set.
+
+    So, for example, one might write:
+
+    class AbstractFileClass:
+        ALLOW_SOFT_DELETE = CustomizableProperty('ALLOW_SOFT_DELETE',
+                                                 description='a boolean saying whether soft delete is allowed')
+        def delete(self, file):
+            if getattr_customized(cls, 'ALLOW_SOFT_DELETE'):
+                self.soft_delete(file)
+            else:
+                self.hard_delete(file)
+
+    Note that there may not be a reasonable default for ALLOW_SOFT_DELETE. Any value would be taken as a boolean.
+    It would be possible to leave the variable unset, but then linters would complain about referring to it.
+    And it would be confusing if accessed without setting it.
+    """
+    # This will raise an error if the attribute is a CustomizableProperty living in the class part of the dict,
+    # but will return the object if it's in the instance.
+    value = getattr(thing, key)
+    if isinstance(value, CustomizableProperty):
+        # This is an uncustomized instance variable, not a class variable.
+        # That's not an intended use case, but just report it without involving mention of the class.
+        raise UncustomizedInstance(instance=thing, field=key)
+    else:
+        return value
